+++ lib/src/cache_stream/response_streams/download_stream.dart
@@ -6,7 +6,7 @@ import 'package:http_cache_stream/src/models/exceptions/invalid_cache_exceptions
 import 'package:http_cache_stream/src/models/http_range/http_range_response.dart';
 
 import '../../etc/extensions/http_extensions.dart';
-import '../../models/config/stream_cache_config.dart';
+import '../../models/cache_config/stream_cache_config.dart';
 import '../../models/exceptions/http_exceptions.dart';
 import '../../models/metadata/cached_response_headers.dart';
 import '../../models/stream_requests/int_range.dart';