+++ lib/src/models/exceptions/state_errors.dart
@@ -1,0 +1,5 @@
+class CacheStreamDisposedException extends StateError {
+  final Uri sourceUrl;
+  CacheStreamDisposedException(this.sourceUrl)
+      : super('Attempted to use a disposed HttpCacheStream | $sourceUrl');
+}