+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.5 
+* Add support for HTTP HEAD requests
 
 * Add `requestTimeout` to cache configuration. Cache requests that are not fulfilled within this duration are closed with a `StreamRequestTimedOutException`. Defaults to 60 seconds.