+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.5 
 * Add support for HTTP HEAD requests
 
-* Add `cacheFileResolver` to global cache configuration
+* Add `cacheFileResolver` to global cache configuration. This function is used to determine the output cache file path when a stream is created.
 
 * Add `requestTimeout` to cache configuration. Cache requests that are not fulfilled within this duration are closed with a `StreamRequestTimedOutException`. Defaults to 60 seconds.